Success in rendering dropdown from Demo instead of Query
