+++ demo/client/src/helper-functions/HelperFunctions.js
@@ -1,10 +1,4 @@
-/**
- * @param {Array} newList - Array of query fields
- * @param {Array} sub - Array of query fields in the sub-query (aka "cities" in "countries")
- * @param {string} query - the query name for ex: "countries" or "cities"
- * @param {string} id - a number when querying by id
- * @param {Object} currentResults - the last output of the function - looks like { countries: ['id', 'name', {'cities': ['id', 'name']}] }
- */
+
 
 /* 
   Used in Query and Demo
@@ -24,128 +18,126 @@
     - looks like: { QUERY: ['item1', 'item2', {'cities': ['item1', 'item2']}] }
 */
 
-const ResultsHelper = (newList, sub, query, id, currentResults) => {
+// const ResultsHelper = (newList, sub, query, id, currentResults) => {
 
-  for (let arr in currentResults) {
-    //===========================//
-    //===Alters the main array===//
-    //===========================//
+//   for (let arr in currentResults) {
+//     //===========================//
+//     //===Alters the main array===//
+//     //===========================//
 
-    if (newList) {
-      const currentList = currentResults[arr];
+//     if (newList) {
+//       const currentList = currentResults[arr];
 
-      // determine whether we already have cities
-      let alreadyHaveCities = false;
-      currentList.forEach((el) => {
-        if (typeof el === 'object') alreadyHaveCities = true;
-      });
+//       // determine whether we already have cities
+//       let alreadyHaveCities = false;
+//       currentList.forEach((el) => {
+//         if (typeof el === 'object') alreadyHaveCities = true;
+//       });
 
-      // determine whether newList has it
-      let newListHasCities = false;
-      newList.forEach((el) => {
-        if (el === 'cities') newListHasCities = true;
-      });
+//       // determine whether newList has it
+//       let newListHasCities = false;
+//       newList.forEach((el) => {
+//         if (el === 'cities') newListHasCities = true;
+//       });
 
-      // if we already have it but new list doesn't, we're deleting it
-      if (alreadyHaveCities === true && newListHasCities === false) {
-        currentList.forEach((el, i) => {
-          if (typeof el === 'object') currentList.splice(i, 1);
-        });
-      }
+//       // if we already have it but new list doesn't, we're deleting it
+//       if (alreadyHaveCities === true && newListHasCities === false) {
+//         currentList.forEach((el, i) => {
+//           if (typeof el === 'object') currentList.splice(i, 1);
+//         });
+//       }
 
-      // if new list has it but we don't, we're adding it with the default initial values
-      if (alreadyHaveCities === false && newListHasCities === true) {
-        currentList.push({ cities: ['id'] });
-      }
+//       // if new list has it but we don't, we're adding it with the default initial values
+//       if (alreadyHaveCities === false && newListHasCities === true) {
+//         currentList.push({ cities: ['id'] });
+//       }
 
-      currentResults[arr] = currentList; // if no cities, this doesn't get altered
+//       currentResults[arr] = currentList; // if no cities, this doesn't get altered
 
-      // if we are NOT DEALING WITH CITIES AT ALL
-      if (alreadyHaveCities === false && newListHasCities === false) {
-        currentResults[arr] = newList;
-      }
+//       // if we are NOT DEALING WITH CITIES AT ALL
+//       if (alreadyHaveCities === false && newListHasCities === false) {
+//         currentResults[arr] = newList;
+//       }
 
-      // if we need to simply preserve cities as it is
-      if (alreadyHaveCities === true && newListHasCities === true) {
-        let storeCityObject;
-        currentList.forEach((el) => {
-          if (typeof el === 'object') storeCityObject = el;
-        });
+//       // if we need to simply preserve cities as it is
+//       if (alreadyHaveCities === true && newListHasCities === true) {
+//         let storeCityObject;
+//         currentList.forEach((el) => {
+//           if (typeof el === 'object') storeCityObject = el;
+//         });
 
-        // loop through newList and
-        const finalList = newList.map((el) => {
-          if (el === 'cities') {
-            return storeCityObject;
-          }
-          return el;
-        });
-        currentResults[arr] = finalList;
-      }
-    }
+//         // loop through newList and
+//         const finalList = newList.map((el) => {
+//           if (el === 'cities') {
+//             return storeCityObject;
+//           }
+//           return el;
+//         });
+//         currentResults[arr] = finalList;
+//       }
+//     }
 
     //===============================//
     //===Alters the city sub-array===//
     //===============================//
 
-    if (sub) {
-      const currentList = currentResults[arr];
-      currentList.forEach((el, i) => {
-        if (typeof el === 'object') {
-          for (let x in el) {
-            currentResults[arr][i][x] = sub;
-          }
-        }
-      });
-    }
+    // if (sub) {
+    //   const currentList = currentResults[arr];
+    //   currentList.forEach((el, i) => {
+    //     if (typeof el === 'object') {
+    //       for (let x in el) {
+    //         currentResults[arr][i][x] = sub;
+    //       }
+    //     }
+    //   });
+    // }
 
     //======================//
     //===Alters the query===//
     //======================//
 
-    if (query) {
-      let fields;
-      if (query === 'country by id') {
-        query = 'country (id:1)';
-        fields = currentResults[arr];
-      }
-      if (query === 'cities by country id') {
-        query = 'citiesByCountry (country_id:1)';
-        fields = currentResults[arr];
-      }
-      if (query === 'countries' || query === 'cities') {
-        fields = ['id'];
-      }
-      currentResults[query] = fields;
-      delete currentResults[arr];
-    }
+    // if (query) {
+    //   let fields;
+    //   if (query === 'country by id') {
+    //     query = 'country (id:1)';
+    //     fields = currentResults[arr];
+    //   }
+    //   if (query === 'cities by country id') {
+    //     query = 'citiesByCountry (country_id:1)';
+    //     fields = currentResults[arr];
+    //   }
+    //   if (query === 'countries' || query === 'cities') {
+    //     fields = ['id'];
+    //   }
+    //   currentResults[query] = fields;
+    //   delete currentResults[arr];
+    // }
 
     //===================//
     //===Alters the id===//
     //===================//
 
-    if (id) {
-      let query = arr;
-      if (query.includes(':')) {
-        let index = query.indexOf(':');
-        query = query.slice(0, index + 1);
-      }
-      query += id + ')';
-      currentResults[query] = currentResults[arr];
-      delete currentResults[arr];
-    }
-  }
+  //   if (id) {
+  //     let query = arr;
+  //     if (query.includes(':')) {
+  //       let index = query.indexOf(':');
+  //       query = query.slice(0, index + 1);
+  //     }
+  //     query += id + ')';
+  //     currentResults[query] = currentResults[arr];
+  //     delete currentResults[arr];
+  //   }
+  // }
 
   // RETURN STATEMENT FOR ALL
-  return currentResults;
-};
+//   return currentResults;
+// };
 
 //======================================//
 //========== CreateQueryStr ============//
 //======================================//
 
-/**
- * @param {Object} currentResults - looks like { countries: ['id', 'name', {'cities': ['id', 'name']}] }
- */
+
 
  function CreateQueryStr(queryObject, operationType) {
   if (Object.keys(queryObject).length === 0) return ''
@@ -212,4 +204,4 @@ const ResultsHelper = (newList, sub, query, id, currentResults) => {
 
 //===============EXPORT=================//
 
-export { ResultsHelper, CreateQueryStr };
+export { CreateQueryStr };
